requirements.txt
torch
numpy
einops
opencv-python
matplotlib
detectron2
Pillow
open_clip